Feat: set theme only in android
