--- composeApp/src/iosMain/kotlin/ui/theme/Theme.ios.kt
@@ -1,13 +1,0 @@
-package ui.theme
-
-import androidx.compose.material3.ColorScheme
-import androidx.compose.runtime.Composable
-import platform.Foundation.NSLog
-import platform.UIKit.UIBarMetrics
-import platform.UIKit.UIColor
-import platform.UIKit.UIImage
-import platform.UIKit.UINavigationBar
-import platform.UIKit.UINavigationController
-
-@Composable
-actual fun setupSystemParams(colorScheme: ColorScheme) {  }
